\name{auc_probability}
\alias{auc_probability}
\title{AUC}
\description{
Compute AUC, or discrimination accuracy, of each potency calculator for distinguishing between most vs. least differentiated populations }
\usage{
auc_probability(labels, scores, N = 1e+07)
}
\arguments{
  \item{labels}{ 
}
  \item{scores}{
}
  \item{N}{
}
}

\value{ numeric }
\references{
https://www.r-bloggers.com/2016/11/calculating-auc-the-area-under-a-roc-curve/}
