\name{g2m_genes_celeg}
\alias{g2m_genes_celeg}
\docType{data}
\title{
G2M cell cycle genes, C elegans
}
\description{
From Scanpy cell cycle scoring function, converted to C. elegans using gprofiler2
}
\usage{data("g2m_genes_celeg")}
\format{
  The format is:
 chr [1:100] "unc-104" "klp-12" "klp-18" "klp-19" "klp-20" "klp-13" "klp-4" ...
}