\name{g2m_genes_mouse}
\alias{g2m_genes_mouse}
\docType{data}
\title{
G2M cell cycle genes, Mus musculus}
\description{
From Scanpy cell cycle scoring function, converted to Mus musculus using gprofiler2}
\usage{data("g2m_genes_mouse")}
\format{
  The format is:
 chr [1:52] "Cbx5" "Gtse1" "Dlgap5" "Smc4" "Cdc25c" "Ctcf" "Kif20b" "Cdca2" "Cdca3" "Hmgb2" "G2e3" "Anp32e" ...
}
