\name{mmTFs}
\alias{mmTFs}
\docType{data}
\title{reference Mus musculus TF list
}
\description{list of known transcription factors in Mus musculus}
\usage{data("mmTFs")}
\format{
  The format is:
 chr [1:1838] "Adnp" "Adnp2" "Aebp1" "Aebp2" "Aff1" "Aff2" "Aff3" "Aff4" "Ago1" "Ago2" "Ahctf1" "Ahr" "Ahrr" ...
}
