\name{s_genes_celeg}
\alias{s_genes_celeg}
\docType{data}
\title{
S cell cycle genes, C. elegans}
\description{Scanpy cell cycle scoring S genes, converted to C. elegans with ggprofiler2}
\usage{data("s_genes_celeg")}
\format{
  The format is:
 chr [1:41] "orc-1" "rtel-1" "chl-1" "xpd-1" "xpg-1" "exo-1" "cye-1" "cyd-1" "cyb-2.1" "cyb-2.2" "cyb-1" "cyb-3" ...
}
