\name{s_genes_mouse}
\alias{s_genes_mouse}
\docType{data}
\title{
S cell cycle genes, Mus musculus
}
\description{
From Scanpy cell cycle scoring function, converted to M. musculus using gprofiler2
}
\usage{data("s_genes_mouse")}
\format{
  The format is:
 chr [1:40] "Rrm2" "Mcm4" "Msh2" "Exo1" "Uhrf1" "Chaf1b" "Gmnn" "Mcm2" ...
}