\name{g2m_genes_human}
\alias{g2m_genes_human}
\docType{data}
\title{
G2M cell cycle genes, human}
\description{
From Scanpy cell cycle scoring function, converted to Homo sapiens using gprofiler2}
\usage{data("g2m_genes_human")}
\format{
  The format is:
 chr [1:55] "E2F8" "HMGB2" "CDK1" "NUSAP1" "UBE2C" "BIRC5" "TPX2" "TOP2A" "NDC80" "CKS2" "NUF2" "CKS1B" "MKI67" ...
}
