\name{s_genes_celeg}
\alias{s_genes_celeg}
\docType{data}
\title{
S cell cycle genes, C. elegans
}
\description{
From Scanpy cell cycle scoring function, converted to C. elegans using gprofiler2
}
\usage{data("s_genes_celeg")}
\format{
  The format is:
 chr [1:41] "orc-1" "rtel-1" "chl-1" "xpd-1" "xpg-1" "exo-1" "cye-1" ...
}