\name{s_genes_human}
\alias{s_genes_human}
\docType{data}
\title{
S cell cycle genes, Homo sapiens
}
\description{
From Scanpy cell cycle scoring function, converted to H. sapiens using gprofiler2
}
\usage{data("s_genes_human")}
\format{
  The format is:
 chr [1:43] "MCM5" "PCNA" "TYMS" "FEN1" "MCM2" "MCM4" "RRM1" "UNG" "GINS2" ...
}