\name{s_genes_mouse}
\alias{s_genes_mouse}
\docType{data}
\title{
S cell cycle genes, Mus musculus}
\description{Scanpy cell cycle scoring S genes}
\usage{data("s_genes_mouse")}
\format{
  The format is:
 chr [1:40] "Rrm2" "Mcm4" "Msh2" "Exo1" "Uhrf1" "Chaf1b" "Gmnn" "Mcm2" "Cdc45" "Fen1" "Slbp" "Dscc1" "Ubr7" ...
}
